+++ code/core/package.json
@@ -280,6 +280,9 @@
       ],
       "internal/instrumenter": [
         "./dist/instrumenter/index.d.ts"
+      ],
+      "test": [
+        "./dist/test/index.d.ts"
       ]
     }
   },
@@ -295,6 +298,9 @@
     "prep": "jiti ./scripts/prep.ts"
   },
   "dependencies": {
+    "@testing-library/dom": "10.4.0",
+    "@testing-library/jest-dom": "6.5.0",
+    "@testing-library/user-event": "14.5.2",
     "better-opn": "^3.0.2",
     "browser-assert": "^1.2.1",
     "esbuild": "^0.18.0 || ^0.19.0 || ^0.20.0 || ^0.21.0 || ^0.22.0 || ^0.23.0 || ^0.24.0 || ^0.25.0",
@@ -349,6 +355,7 @@
     "@types/react-transition-group": "^4",
     "@types/semver": "^7.5.8",
     "@types/ws": "^8",
+    "@vitest/spy": "^2.0.5",
     "@vitest/utils": "^2.1.1",
     "@yarnpkg/esbuild-plugin-pnp": "^3.0.0-rc.10",
     "@yarnpkg/fslib": "2.10.3",
@@ -359,6 +366,7 @@
     "boxen": "^7.1.1",
     "browser-dtector": "^3.4.0",
     "camelcase": "^8.0.0",
+    "chai": "^5.1.1",
     "cli-table3": "^0.6.1",
     "commander": "^12.1.0",
     "comment-parser": "^1.4.1",
@@ -423,7 +431,7 @@
     "strip-json-comments": "^5.0.1",
     "telejson": "^7.2.0",
     "tiny-invariant": "^1.3.1",
-    "tinyspy": "^2.2.0",
+    "tinyspy": "^3.0.2",
     "ts-dedent": "^2.0.0",
     "tsconfig-paths": "^4.2.0",
     "type-fest": "^4.18.1",
