+++ code/core/src/test/spy.test.ts
