+++ code/core/src/test/spy.ts
