+++ code/core/template/stories/destructuring-not-transpiled.stories.ts
