+++ code/core/template/stories/module-mocking.stories.ts
