+++ code/core/template/stories/mount-in-play.stories.ts
