--- code/lib/test/package.json
@@ -1,84 +1,0 @@
-{
-  "name": "@storybook/test",
-  "version": "9.0.0-alpha.1",
-  "description": "",
-  "keywords": [
-    "storybook"
-  ],
-  "homepage": "https://github.com/storybookjs/storybook/tree/next/code/lib/test",
-  "bugs": {
-    "url": "https://github.com/storybookjs/storybook/issues"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/storybookjs/storybook.git",
-    "directory": "code/lib/test"
-  },
-  "funding": {
-    "type": "opencollective",
-    "url": "https://opencollective.com/storybook"
-  },
-  "license": "MIT",
-  "sideEffects": true,
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "node": "./dist/index.js",
-      "import": "./dist/index.mjs",
-      "require": "./dist/index.js"
-    },
-    "./package.json": "./package.json"
-  },
-  "main": "dist/index.js",
-  "module": "dist/index.mjs",
-  "types": "dist/index.d.ts",
-  "files": [
-    "dist/**/*",
-    "README.md",
-    "*.js",
-    "*.d.ts"
-  ],
-  "scripts": {
-    "check": "jiti ../../../scripts/prepare/check.ts",
-    "prep": "jiti ../../../scripts/prepare/bundle.ts"
-  },
-  "dependencies": {
-    "@storybook/global": "^5.0.0",
-    "@testing-library/dom": "10.4.0",
-    "@testing-library/jest-dom": "6.5.0",
-    "@testing-library/user-event": "14.5.2",
-    "@vitest/expect": "2.0.5",
-    "@vitest/spy": "2.0.5"
-  },
-  "devDependencies": {
-    "chai": "^5.1.1",
-    "tinyspy": "^3.0.0",
-    "ts-dedent": "^2.2.0",
-    "type-fest": "~2.19",
-    "typescript": "^5.7.3"
-  },
-  "peerDependencies": {
-    "storybook": "workspace:^"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "bundler": {
-    "entries": [
-      "./src/index.ts"
-    ],
-    "noExternal": [
-      "@testing-library/dom",
-      "@testing-library/jest-dom",
-      "@testing-library/user-event",
-      "chai",
-      "@vitest/expect",
-      "@vitest/spy",
-      "@vitest/utils"
-    ],
-    "externals": [
-      "util"
-    ]
-  },
-  "gitHead": "e6a7fd8a655c69780bc20b9749c2699e44beae16"
-}
