--- code/lib/test/project.json
@@ -1,8 +1,0 @@
-{
-  "name": "test",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "projectType": "library",
-  "targets": {
-    "build": {}
-  }
-}
