--- code/lib/test/template/stories/utils.mock.ts
@@ -1,5 +1,0 @@
-import { fn } from '@storybook/test';
-
-import * as utils from './utils.ts';
-
-export const foo = fn(utils.foo).mockName('foo');
