--- code/lib/test/template/stories/utils.ts
@@ -1,1 +1,0 @@
-export const foo = () => 'not mocked';
