--- code/lib/test/vitest.config.ts
@@ -1,10 +1,0 @@
-import { defineConfig, mergeConfig } from 'vitest/config';
-
-import { vitestCommonConfig } from '../../vitest.workspace';
-
-export default mergeConfig(
-  vitestCommonConfig,
-  defineConfig({
-    // Add custom config here
-  })
-);
